# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tinder API Documentation",
        "description": "API for Tinder-like dating application",
        "contact": {
            "email": "support@tinder-api.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/people/recommended": {
            "get": {
                "tags": [
                    "People"
                ],
                "summary": "Get list of recommended people",
                "operationId": "01ebfac93484bfc05152fc6c15421433",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Current user ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/people/{id}/like": {
            "post": {
                "tags": [
                    "People"
                ],
                "summary": "Like a person",
                "operationId": "9ca1b73b0ea8c9be75705e22990243a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Target user ID to like",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Current user ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Person liked successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Person liked successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Already swiped this person"
                    }
                }
            }
        },
        "/people/{id}/dislike": {
            "post": {
                "tags": [
                    "People"
                ],
                "summary": "Dislike a person",
                "operationId": "079025c24928c1f1071f6f71417911a8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Target user ID to dislike",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Current user ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Person disliked successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Person disliked successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Already swiped this person"
                    }
                }
            }
        },
        "/people/liked": {
            "get": {
                "tags": [
                    "People"
                ],
                "summary": "Get list of liked people",
                "operationId": "4ba543911d542e18c853a08ad27ed437",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Current user ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "People",
            "description": "People"
        }
    ]
}